exercise2_bcra_api/requirements.txt
# exercise2_bcra_api/requirements.txt
requests
pandas
sqlalchemy
psycopg2-binary
python-dotenv